requirements.txt
asgiref==3.2.7
Django==3.0.7
django-ckeditor==5.9.0
django-filter==2.3.0
django-js-asset==1.2.2
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
modeltranslation==0.25
Pillow==7.1.2
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
